+++ src/actions/get_card_content.rs
@@ -1,0 +1,1 @@
+pub fn call() {}
