+++ src/actions/mod.rs
@@ -1,8 +1,6 @@
-pub mod get_jira_config;
-pub mod get_jira_pr_title;
-pub mod get_jira_title;
+pub mod get_pr_title;
+pub mod get_card_title;
 pub mod progress_card;
 pub mod review_card;
-pub mod transitions;
 pub mod assignee_card;
-pub mod create_new_branch;
+pub mod get_card_content;
