+++ src/actions/progress_card.rs
@@ -2,7 +2,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::actions;
 use crate::jira_api;
-use super::transitions::TRANSITIONS;
+use crate::git_api;
+use crate::jira_api::transitions::TRANSITIONS;
 
 #[derive(Deserialize, Serialize)]
 struct Transition {
@@ -14,7 +15,7 @@ pub fn call(code: &usize) -> Result<bool, Box<dyn std::error::Error>> {
 
     let assignee_response = actions::assignee_card::call(code).unwrap();
 
-    let branch_response = actions::create_new_branch::call(code).unwrap();
+    let branch_response = git_api::create_branch::call(code).unwrap();
 
     let all_good = transition_response.status_code == 200
         && assignee_response.status_code == 200
