+++ src/actions/review_card.rs
@@ -1,4 +1,4 @@
-use crate::actions::transitions::TRANSITIONS;
+use crate::jira_api::transitions::TRANSITIONS;
 use crate::git_api;
 use crate::{actions, jira_api};
 use std::error;
@@ -8,7 +8,7 @@ pub fn call(code: &usize) -> Result<bool, Box<dyn error::Error>> {
         jira_api::move_card_status::call(code, &TRANSITIONS.review).unwrap();
     let branches = git_api::list_branches::call().unwrap();
     let has_develop = branches.iter().any(|s| s == "develop");
-    let pr_title = actions::get_jira_pr_title::call(code).unwrap();
+    let pr_title = actions::get_pr_title::call(code).unwrap();
 
     if transition_response.status_code == 204 {
         let base_branch = if has_develop { "develop" } else { "main" };
