+++ src/jira_api/get_card_content.rs
@@ -1,0 +1,2 @@
+pub fn call() {
+}
