+++ src/jira_api/mod.rs
@@ -1,1 +1,4 @@
 pub mod move_card_status;
+pub mod get_card_content;
+pub mod get_config;
+pub mod transitions;
